requirements.txt
asgiref==3.2.10
backcall==0.2.0
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
Django==3.0.7
entrypoints==0.3
fuzzywuzzy==0.18.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gspread==3.6.0
httplib2==0.18.1
idna==2.9
IMDbPY==6.8
jedi==0.17.0
lxml==4.5.1
oauth2client==4.1.3
oauthlib==3.1.0
parso==0.7.0
pick==0.6.7
pickleshare==0.7.5
prompt-toolkit==3.0.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
pywin32==227
pyzmq==18.1.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.4
wincertstore==0.2
windows-curses==2.1.0
